pyproject.toml
[tool.poetry]
name = "gpt-plus"
version = "0.1.0"
description = ""
authors = ["Ozel Yilmazel"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
openai = "^1.12.0"
pyautogen = "^0.2.14"
tiktoken = "^0.6.0"
pytest = "^8.0.1"
load-dotenv = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
